requirements.txt
# Core Framework
streamlit>=1.30.0

# Azure Services
azure-search-documents>=11.4.0
azure-storage-blob>=12.17.0
azure-identity>=1.13.0
azure-core>=1.29.5

# LangChain & Related
langchain>=0.1.0
langchain-openai>=0.0.2
langchain-core>=0.1.10
langchain-community>=0.0.13
langgraph>=0.0.10

# Document Processing
pypdf>=3.17.1
python-docx>=1.0.1
python-magic-bin>=0.4.14  # For Windows
python-magic>=0.4.27     # For Unix-like systems

# OpenAI
openai>=1.10.0

# Web Search
google-search-results>=2.4.2  # SerpAPI

# Utilities
python-dotenv>=1.0.0
typing-extensions>=4.8.0
aiohttp>=3.9.1
certifi>=2023.11.17
charset-normalizer>=3.3.2
idna>=3.6
urllib3>=2.1.0

# Optional: Development Tools
pylint>=3.0.3
black>=24.1.1
pytest>=7.4.4

